// src/app/interfaces/filter-config.interface.ts
export interface FilterConfig {
    storeId: string;
    categoryId: string;
    productId: string;
    brandId: string;
    genreId: string;
    userId: string;
    method: string;
    rateId: string;
    toDate: string;
    fromDate: string;
    selectedPeriod: string;
    searchedProduct: string;
}